# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Crypto Trader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /broker/connect:
    post:
      tags: [broker]
      operationId: connectBroker
      summary: Link an Alpaca brokerage account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectBrokerInput"
      responses:
        "200":
          description: Broker connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
  /broker/status:
    get:
      tags: [broker]
      operationId: getBrokerStatus
      summary: Get broker connection status and account summary
      responses:
        "200":
          description: Broker status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerStatus"
        "401":
          description: Unauthorized
  /broker/disconnect:
    delete:
      tags: [broker]
      operationId: disconnectBroker
      summary: Remove linked brokerage account
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
        "401":
          description: Unauthorized
  /broker/positions:
    get:
      tags: [broker]
      operationId: listBrokerPositions
      summary: List open positions in the linked brokerage account
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerPosition"
        "400":
          description: No broker connected
        "401":
          description: Unauthorized
  /broker/orders:
    get:
      tags: [broker]
      operationId: listBrokerOrders
      summary: List recent orders in the linked brokerage account
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerOrder"
        "400":
          description: No broker connected
        "401":
          description: Unauthorized
    post:
      tags: [broker]
      operationId: placeBrokerOrder
      summary: Place a real-money market order (notional, fractional supported)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBrokerOrderInput"
      responses:
        "200":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerOrder"
        "400":
          description: Invalid order or broker error
        "401":
          description: Unauthorized
  /broker/auto-trade:
    get:
      tags: [broker]
      operationId: getBrokerAutoTradeSettings
      summary: Get real-money auto-trade settings for the linked broker account
      responses:
        "200":
          description: Real-money auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAutoTradeSettings"
        "400":
          description: No broker connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateBrokerAutoTradeSettings
      summary: Update real-money auto-trade settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAutoTradeSettings"
        "400":
          description: No broker connected or invalid input
        "401":
          description: Unauthorized
  /broker/oanda/connect:
    post:
      tags: [broker]
      operationId: connectOanda
      summary: Link an OANDA account for forex, commodity, and index CFD trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectOandaInput"
      responses:
        "200":
          description: OANDA connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
  /broker/oanda/status:
    get:
      tags: [broker]
      operationId: getOandaStatus
      summary: Get OANDA connection status and account summary
      responses:
        "200":
          description: OANDA status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaStatus"
        "401":
          description: Unauthorized
  /broker/oanda/disconnect:
    delete:
      tags: [broker]
      operationId: disconnectOanda
      summary: Remove linked OANDA account
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
  /broker/oanda/auto-trade:
    get:
      tags: [broker]
      operationId: getOandaAutoTradeSettings
      summary: Get real-money auto-trade settings for the linked OANDA account
      responses:
        "200":
          description: OANDA auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAutoTradeSettings"
        "400":
          description: No OANDA account connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateOandaAutoTradeSettings
      summary: Update real-money auto-trade settings for OANDA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAutoTradeSettings"
        "400":
          description: No OANDA account connected or invalid input
        "401":
          description: Unauthorized
  /broker/kraken/connect:
    post:
      tags: [broker]
      operationId: connectKraken
      summary: Link a Kraken Spot account for manual crypto trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectKrakenInput"
      responses:
        "200":
          description: Kraken connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
  /broker/kraken/status:
    get:
      tags: [broker]
      operationId: getKrakenStatus
      summary: Get Kraken Spot connection and balance status
      responses:
        "200":
          description: Kraken connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenStatus"
        "401":
          description: Unauthorized
  /broker/kraken/disconnect:
    delete:
      tags: [broker]
      operationId: disconnectKraken
      summary: Remove linked Kraken Spot account
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
  /broker/kraken/auto-trade:
    get:
      tags: [broker]
      operationId: getKrakenAutoTradeSettings
      summary: Get Kraken AI auto-trade limits
      responses:
        "200":
          description: Kraken AI auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenAutoTradeSettings"
        "400":
          description: No Kraken account connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateKrakenAutoTradeSettings
      summary: Update Kraken AI auto-trade settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated Kraken AI auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenAutoTradeSettings"
        "400":
          description: No Kraken account connected or invalid settings
        "401":
          description: Unauthorized
components:
...
  schemas:
...
    ConnectBrokerInput:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 1
        apiSecret:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [paper, live]
      required: [apiKey, apiSecret, mode]
    BrokerStatus:
      type: object
      properties:
        connected:
          type: boolean
        mode:
          type: string
          enum: [paper, live]
        apiKeyMasked:
          type: string
        accountNumber:
          type: string
        currency:
          type: string
        cash:
          type: number
        buyingPower:
          type: number
        equity:
          type: number
        portfolioValue:
          type: number
        status:
          type: string
        longMarketValue:
          type: number
        shortMarketValue:
          type: number
        daytradeCount:
          type: number
        patternDayTrader:
          type: boolean
        shortingEnabled:
          type: boolean
        multiplier:
          type: number
        maintenanceMargin:
          type: number
        initialMargin:
          type: number
        lastEquity:
          type: number
        createdAt:
          type: string
      required: [connected]
    BrokerPosition:
      type: object
      properties:
        symbol:
          type: string
        assetClass:
          type: string
        qty:
          type: number
        marketValue:
          type: number
        avgEntryPrice:
          type: number
        currentPrice:
          type: number
        unrealizedPl:
          type: number
        unrealizedPlPercent:
          type: number
        side:
          type: string
      required:
        [
          symbol,
          assetClass,
          qty,
          marketValue,
          avgEntryPrice,
          currentPrice,
          unrealizedPl,
          unrealizedPlPercent,
          side,
        ]
    PlaceBrokerOrderInput:
      type: object
      properties:
        symbol:
          type: string
          minLength: 1
        side:
          type: string
          enum: [buy, sell]
        notional:
          type: number
          minimum: 1
      required: [symbol, side, notional]
    BrokerAutoTradeSettings:
      type: object
      properties:
        enabled:
          type: boolean
        maxPerTradeUsd:
          type: number
        dailyLimitUsd:
          type: number
        spentTodayUsd:
          type: number
        mode:
          type: string
          enum: [paper, live]
      required: [enabled, maxPerTradeUsd, dailyLimitUsd, spentTodayUsd, mode]
...
    BrokerOrder:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
        notional:
          type: number
          nullable: true
        qty:
          type: number
          nullable: true
        filledQty:
          type: number
          nullable: true
        filledAvgPrice:
          type: number
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required: [id, symbol, side, status, createdAt]
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean